moves TemporalValue to a more loved place (down to data lvl)
