+++ history.md
@@ -13,8 +13,8 @@
 - New class `meico.mei.ornament.OrnamentProcessor` to integrate the ornament expansion into the MEI to MSM/MPM conversion process (`checkIfOrnament()`, `processOrnament()`, `addToOrnamentationMap()`), with helper methods (`flattenGraceGrp()`, `MeiNote2MsmNote()`, `getRptString()`). `processOrnament()` splits combined ornament expansions (multiple `graceGrp` segments) into separate `OrnamentData` entries, each with its own `ornamentDefName`, `at end`, and `groupId`.
 - New interface `meico.mei.Context` for accessing current context data in processors.
 - New member `ornamentProcessor` in class `meico.mei.Mei2MsmMpmConverter` to delegate ornament specific processing.
-- New class `meico.mpm.elements.TemporalValue` for temporal domain handling (milliseconds, ticks, prepares for note length, relative) with some basic conversion utilities.
-- Enhances class `meico.mpm.elements.styles.defs.OrnamentDef` with the use of `meico.mpm.elements.TemporalValue` for its temporal attributes. Added some standard temporal spreads for ornaments (including `"grace acc delayed"`, `"grace unacc delayed"` defaults). 
+- New class `meico.mpm.elements.maps.data.TemporalValue` for temporal domain handling (milliseconds, ticks, prepares for note length, relative) with some basic conversion utilities.
+- Enhances class `meico.mpm.elements.styles.defs.OrnamentDef` with the use of `meico.mpm.elements.maps.data.TemporalValue` for its temporal attributes. Added some standard temporal spreads for ornaments (including `"grace acc delayed"`, `"grace unacc delayed"` defaults). 
 - New attribute for `TemporalSpread` for defining the `alignment` of an ornament that is either `at start` or `at end`, with providing a method `isAtEnd()`: when true (so `alignment` is `at end`, the ornament is placed at the end of the principal note. `TemporalSpread.apply()` now accepts optional `effectiveFrameStart`/`effectiveFrameLength` overrides for multi-ornament proportional distribution and handles `at end` during rendering.
 - New methods in class `meico.mpm.elements.maps.OrnamentationMap` to handle the new ornamentation data and render it to the map by applying ornament notes (`applyNotesToMaps()`, `copyNotePerfInformation()`) in `apply()`. Added support for multiple OrnamentDefs per principal note: if total frameLengths exceeds 100% of the principal note, the ornaments are proportionally distributed across the note's duration. New helper methods `resolveFrameValues()`, `getPrincipalDuration()`, and inner class `OrnamentEntry`.
