+++ src/meico/mpm/elements/styles/defs/OrnamentDef.java
@@ -2,7 +2,7 @@ package meico.mpm.elements.styles.defs;
 
 import meico.mei.Helper;
 import meico.mpm.Mpm;
-import meico.mpm.elements.TemporalValue;
+import meico.mpm.elements.maps.data.TemporalValue;
 import meico.msm.MsmNoteElement;
 import meico.supplementary.KeyValue;
 import nu.xom.Attribute;
